+++ manual_tests_qp.py
@@ -4,32 +4,29 @@
 from htag import Tag
 
 class App(Tag.body):
-    statics=b"var SESID='no'"
+
     def init(self,param="nada"):
-        self["body"]="background:#EEE;"
+        self["style"]="background:#EEE;"
         self <= Tag.h3("param = "+param)
-        self <= Tag.a("test SELF ?",_href="/",_style="display:block")
-        self <= Tag.a("test SELF pablo",_href="/?pablo",_style="display:block")
-        self <= Tag.button("yo", _onclick=self.test)    # just to control interact
-        self <= Tag.button("quit", _onclick=lambda o: self.exit())    # just to control interact
+        self <= Tag.a("test '/' ?",_href="/",_style="display:block")
+        self <= Tag.a("test '/?param=yo'",_href="/?param=yo",_style="display:block")
+        self <= Tag.button("add content", _onclick=self.add_content)    # just to control interact
+        self <= Tag.button("quit app", _onclick=lambda o: self.exit())    # just to control interact
         self <= Tag.hr()
-        self <= Tag.h3("Only if multiples (WebServer) :")
-        self <= Tag.a("test ANOTHER ?",_href="/toto",_style="display:block")
-        self <= Tag.a("test ANOTHER pablo",_href="/toto?pablo",_style="display:block")
 
+        self <= Tag.h3("Only if it handles tha '/other' route (WebHTTP/DevApp) :")
+        self <= Tag.a("test '/other'"      ,_href="/other",_style="display:block")
+        self <= Tag.a("test '/other?pablo'",_href="/other?pablo",_style="display:block")
+        self <= Tag.hr()
 
-        self( self.bind.rendu( b"SESID" ) )
-
-    def test(self,o):
+    def add_content(self,o):
         self <= "X "
 
-    def rendu(self,x):
-        self <= "SESID:"+x
 
 class AnotherApp(Tag.body):
 
     def init(self, name="vide"):
-        self["body"]="background:#FAA;"
+        self["style"]="background:#FFE;"
         self <= "Hello "+name
         self <= Tag.button("yo", _onclick=self.test)
 
@@ -46,17 +43,18 @@ class AnotherApp(Tag.body):
 # from htag.runners import BrowserStarletteHTTP as Runner
 # from htag.runners import BrowserTornadoHTTP as Runner
 # from htag.runners import ChromeApp as Runner
-from htag.runners import AndroidApp as Runner
+# from htag.runners import AndroidApp as Runner
+# from htag.runners import PyWebView as Runner  # just the "add content" will work (no query params / no other route)
 
 
 app=Runner(App)
 
 if hasattr(app,"serve"): # only DevApp & WebHTTP
-    from starlette.responses import HTMLResponse
+
     def another(request):
         return app.serve(request, AnotherApp )
 
-    app.add_route("/toto", another )
+    app.add_route("/other", another )
 
 
 if __name__=="__main__":